Move PNG files to output folder.
git-svn-id: http://browsershots.googlecode.com/svn/trunk/shotfactory@2190 1175509c-980b-11dd-9476-71c886a649bc

+++ shotfactory.py
@@ -104,7 +104,8 @@ def browsershot(options, server, config, password):
 
     # Upload PNG file
     server.upload_png(config, pngfilename)
-    os.remove(pngfilename)
+    if os.path.exists(pngfilename):
+        os.remove(pngfilename)
 
 
 def error_sleep(message):
